+++ Demo/TiledDemoController.swift
@@ -1516,7 +1516,7 @@ extension TiledDemoController {
         let defaults = UserDefaults.shared
         
         if (defaults.value(forKey: "tiled-gbl-ddoptions") != nil) {
-            self.debugDrawOptions = DebugDrawOptions(rawValue: defaults.integer(forKey: "tiled-gbl-ddoptions"))
+            self.debugDrawOptions = DebugDrawOptions(rawValue: UInt8(defaults.integer(forKey: "tiled-gbl-ddoptions")))
         }
     }
